refactor: extract fuzzy search into a dependency-free core module
The worker previously imported fuzzySearch.ts, which pulls in capture.js
and through it server.ts — booting the entire app module graph per fuzzy
search. That cost ~275ms per call on small files and fired the search
telemetry from inside the worker, where the client identity is never
initialized.

fuzzySearchCore.ts now holds the pure search functions with
fastest-levenshtein as its only import; the worker loads just that and
returns timing metrics as data, which the main thread captures with the
real client id (same event names and payloads as before). Worker
round-trip drops to ~18ms. The worker snippet also gained a .catch so
import/search failures surface as a structured error instead of an
opaque worker crash.
